updated version to `32.0.0`

+++ sstring/prototype_err.h
@@ -4,7 +4,7 @@
 * Commit to this repository at https://github.com/Dark-CodeX/SafeString.git
 * You can use this header file. Do not modify it locally, instead commit it on https://www.github.com
 * File: "prototype_err.h" under "sstring" directory
-* sstring: version 31.5.0
+* sstring: version 32.0.0
 * 
 * MIT License
 * 
@@ -114,7 +114,7 @@ int _is_ascii(sstring *a);
 int _is_alphabetic(sstring *a);
 int _format_escape_sequence(sstring *a);
 
-void __advance__iter_sstring(iter_sstring *is, SIZE_T move_by);
+void __advance__iter_sstring(iter_sstring *is, signed long long int move_by);
 int __c_loop__iter_sstring(iter_sstring *is);
 
 int strcmp_void(const void *a1, const void *a2);
